PaymentService/Core/Application/PaymentProcessorFactory.cs

using Application.MercadoPago;
using Application.Payment.Enums;
using Application.Payment.Ports;
using PaymentsApplication;

namespace Application.Payment
{
    public class PaymentProcessorFactory : IPaymentProcessorFactory
    {
        public IPaymentProcessor GetPaymentProcessor(SupportedPaymentProviders selectedPaymentProvider)
        {
            switch (selectedPaymentProvider)
            {
                case SupportedPaymentProviders.MercadoPago:
                    return new MercadoPagoAdapter();

                default: return new NotImplementedPaymentProvider();
            }
        }
    }
}
